Avoid unnecessary casting, use PDO::PARAM_X more cautiously (#5086)
* maintenance/improvement
* fix userid casting to int
  never use bindParam with an integer variable without PDO::PARAM_INT or the variable will be casted to string,
  better use bindValue to avoid this confusion
* remove bindValue()/bindParam() default parameter PDO::PARAM_STR where possible
* less casting
* move getNextCustomId to AbstractConcreteEntity
* add some missing use statements

* use more arrow functions

+++ src/commands/ImportUser.php
@@ -50,7 +50,7 @@ class ImportUser extends Command
         $userid = (int) $input->getArgument('userid');
         $filePath = $this->Fs->getPath((string) $input->getArgument('file'));
         $uploadedFile = new UploadedFile($filePath, 'input.eln', null, null, true);
-        $teamid = (int) (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
+        $teamid = (new UsersHelper($userid))->getTeamsFromUserid()[0]['id'];
         $Eln = new Eln(new Users($userid, $teamid), sprintf('experiments:%d', $userid), BasePermissions::User->toJson(), BasePermissions::User->toJson(), $uploadedFile, Storage::CACHE->getStorage()->getFs());
         $Eln->import();
